+++ MongoMigrationWebApp/Components/MigrationDetails.razor
@@ -54,6 +54,15 @@
                     </select>
                 </div>
                 <div class="mb-3">
+                    <label for="change-stream-scope" class="form-label" title="Specify the scope of change stream monitoring during online migration.">
+                        Change Stream Scope
+                    </label>                
+                    <select id="change-stream-scope" value="@SelectedChangeStreamScope" class="form-select" @onchange="OnChangeStreamScopeChanged" disabled="@(!NewMode || !isOnline || jobType == JobType.RUOptimizedCopy)">
+                        <option value="Collection" title="Monitor changes at the collection level. Use when collect count is less than 200">Collection</option>
+                        <option value="Server" title="Monitor changes at the server level across all databases. Recommended when migration hundreds of collections.">Server</option>
+                    </select>
+                </div>
+                <div class="mb-3">
                     <label for="change-stream-mode" class="form-label" title="Configure when change stream processing begins during migration.">
                         Change Stream Mode
                     </label>                
@@ -154,6 +163,8 @@
     private string targetEndpoint = string.Empty;   // Target endpoint extracted from the connection string
     private string SelectedMode = "Offline"; // default or loaded value
     private string SelectedChangeStreamMode = "Delayed"; // Default change stream mode
+    private string SelectedChangeStreamScope = "Collection"; // Default change stream type
+    private ChangeStreamLevel changeStreamLevel = ChangeStreamLevel.Collection; // Default to Collection
     private string sourceConnectionString { get; set; } = string.Empty;
     private string targetConnectionString { get; set; } = string.Empty;
     private bool AllowMongoDump => Configuration.GetValue<bool>("AllowMongoDump");
@@ -162,6 +173,13 @@
     {
         var val = e?.Value?.ToString() ?? string.Empty;
         isOnline = val == "Online";
+    }
+
+    private void OnChangeStreamScopeChanged(ChangeEventArgs e)
+    {
+        var val = e?.Value?.ToString() ?? "Collection";
+        SelectedChangeStreamScope = val;
+        changeStreamLevel = val == "Server" ? ChangeStreamLevel.Server : ChangeStreamLevel.Collection;
     }
 
     private void OnChangeStreamModeChanged(ChangeEventArgs e)
@@ -207,8 +225,10 @@
         {
             skipIndexes = true; // Skip indexes for RU optimized copy
             SelectedChangeStreamMode = "Immediate"; // Set to Immediate for RU optimized copy
+            SelectedChangeStreamScope = "Collection"; // Set to Collection for RU optimized copy
             startCSAfterUploads = false;
             aggressiveCS = false;
+            changeStreamLevel = ChangeStreamLevel.Collection; // Force to Collection level
         }       
     }
 
@@ -230,6 +250,8 @@
             aggressiveCS = false;
             startCSAfterUploads = true;
             SelectedChangeStreamMode = "Delayed"; // Default to Delayed
+            SelectedChangeStreamScope = "Collection"; // Default to Collection
+            changeStreamLevel = ChangeStreamLevel.Collection; // Default to Collection
         }
         else
         {
@@ -253,6 +275,7 @@
             isSimulatedRun = Job?.IsSimulatedRun ?? false;
             skipIndexes = Job?.SkipIndexes ?? false;
             appendMode = Job?.AppendMode ?? false;
+            changeStreamLevel = Job?.ChangeStreamLevel ?? ChangeStreamLevel.Collection;
 
             // Set dropdown value based on existing boolean flags
             if (aggressiveCS)
@@ -267,6 +290,18 @@
             {
                 SelectedChangeStreamMode = "Immediate";
             }
+
+            if(changeStreamLevel == ChangeStreamLevel.Collection)
+            {
+                SelectedChangeStreamScope = "Collection";
+            }
+            else
+            {
+                SelectedChangeStreamScope = "Server";
+            }
+
+            // Set change stream type dropdown value based on enum value
+            SelectedChangeStreamScope = changeStreamLevel == ChangeStreamLevel.Server ? "Server" : "Collection";
 
             if(isOnline)
             {
@@ -364,7 +399,8 @@
             AppendMode = appendMode,
             SyncBackEnabled = (enableSyncBack && isOnline && !isSimulatedRun) ? true : false,
             CSStartsAfterAllUploads = startCSAfterUploads,
-            AggresiveChangeStream = aggressiveCS
+            AggresiveChangeStream = aggressiveCS,
+            ChangeStreamLevel = changeStreamLevel
         };
         await OnSubmit.InvokeAsync(job);
     }
