+++ OnlineMongoMigrationProcessor/Models/ChangeStreamLevels.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OnlineMongoMigrationProcessor.Models
+{
+    public enum ChangeStreamLevel
+    {
+        Collection = 0,
+        Server = 1
+    }
+}
